group images by 64-bit distros/older/images
