+++ src/features/Images/ImageSelect.test.tsx
@@ -1,15 +1,15 @@
 import { shallow } from 'enzyme';
 import * as React from 'react';
-
-import { images } from 'src/__data__/images';
-import Select from 'src/components/EnhancedSelect/Select';
-
 import {
-  getImagesOptions,
-  groupImages,
-  groupNameMap,
-  ImageSelect
-} from './ImageSelect';
+  deletedImage1,
+  images,
+  privateImage1,
+  recommendedImage1,
+  recommendedImage2
+} from 'src/__data__/images';
+import Select from 'src/components/EnhancedSelect/Select';
+import { groupNameMap } from 'src/utilities/images';
+import { getImagesOptions, ImageSelect } from './ImageSelect';
 
 const props = {
   classes: { root: '', icon: '', selectContainer: '' },
@@ -19,161 +19,7 @@ const props = {
 
 const component = shallow(<ImageSelect {...props} />);
 
-const privateImage1 = {
-  deprecated: false,
-  type: 'manual',
-  id: 'private/0001',
-  created_by: 'somefella',
-
-  description: '',
-  vendor: null,
-  size: 0,
-  label: '',
-  is_public: false,
-  created: ''
-};
-
-const privateImage2 = {
-  deprecated: false,
-  type: 'manual',
-  id: 'private/0002',
-  created_by: 'somefella',
-
-  description: '',
-  vendor: null,
-  size: 0,
-  label: '',
-  is_public: false,
-  created: ''
-};
-
-const deprecatedImage1 = {
-  deprecated: true,
-  type: 'manual',
-  id: 'linode/0001',
-  created_by: 'linode',
-
-  description: '',
-  vendor: null,
-  size: 0,
-  label: '',
-  is_public: false,
-  created: ''
-};
-
-const deprecatedImage2 = {
-  deprecated: true,
-  type: 'manual',
-  id: 'linode/0002',
-  created_by: 'linode',
-
-  description: '',
-  vendor: null,
-  size: 0,
-  label: '',
-  is_public: false,
-  created: ''
-};
-
-const recommendedImage1 = {
-  deprecated: false,
-  type: 'manual',
-  id: 'linode/0001',
-  created_by: 'linode',
-
-  description: '',
-  vendor: null,
-  size: 0,
-  label: '',
-  is_public: false,
-  created: ''
-};
-
-const recommendedImage2 = {
-  deprecated: false,
-  type: 'manual',
-  id: 'linode/0002',
-  created_by: 'linode',
-
-  description: '',
-  vendor: null,
-  size: 0,
-  label: '',
-  is_public: false,
-  created: ''
-};
-
-const deletedImage1 = {
-  deprecated: false,
-  type: 'automatic',
-  id: 'private/0001',
-  created_by: null,
-
-  description: '',
-  vendor: null,
-  size: 0,
-  label: '',
-  is_public: false,
-  created: ''
-};
-
-const deletedImage2 = {
-  deprecated: false,
-  type: 'automatic',
-  id: 'private/0002',
-  created_by: null,
-
-  description: '',
-  vendor: null,
-  size: 0,
-  label: '',
-  is_public: false,
-  created: ''
-};
-
 describe('ImageSelect', () => {
-  describe('groupImages method', () => {
-    it("should return a group of the user's private images", () => {
-      const result = groupImages([privateImage1, privateImage2]);
-
-      const expected = {
-        images: [privateImage1, privateImage2]
-      };
-
-      expect(result).toEqual(expected);
-    });
-
-    it('should return group deprecated images', () => {
-      const result = groupImages([deprecatedImage1, deprecatedImage2]);
-
-      const expected = {
-        older: [deprecatedImage1, deprecatedImage2]
-      };
-
-      expect(result).toEqual(expected);
-    });
-
-    it('should return group recommended images', () => {
-      const _images = [recommendedImage1, recommendedImage2];
-      const result = groupImages(_images);
-
-      const expected = {
-        recommended: _images
-      };
-
-      expect(result).toEqual(expected);
-    });
-
-    it('should return group deleted images', () => {
-      const result = groupImages([deletedImage1, deletedImage2]);
-
-      const expected = {
-        deleted: [deletedImage1, deletedImage2]
-      };
-
-      expect(result).toEqual(expected);
-    });
-  });
   describe('getImagesOptions function', () => {
     it('should return a list of GroupType', () => {
       const items = getImagesOptions([recommendedImage1, recommendedImage2]);
